[include] Fix a bad return type for ComHeight (float instead of double).
	* include/walkGenJrl/ZMPRefTrajectoryGeneration/ZMPRefTrajectoryGeneration.h: Fixed a pb regarding type returned by ComHeight().

+++ include/walkGenJrl/ZMPRefTrajectoryGeneration/ZMPRefTrajectoryGeneration.h
@@ -120,7 +120,7 @@ namespace PatternGeneratorJRL
       { return m_Omega;};
 
     /* ! \brief Returns the Com Height. */
-    inline const float & GetComHeight() const
+    inline const double & GetComHeight() const
     { return m_ComHeight; };
     
     /* ! \brief Returns the Com Height. */
